requirements.txt
Flask==2.3.2
numpy==1.24.3
opencv_python==4.7.0.72
paho_mqtt==1.6.1
PyYAML==6.0.1
tflite_support==0.4.3
requests==2.30.0
Pillow==9.5.0
prometheus_client==0.17.1
prometheus-flask-exporter==0.22.4
